fix: refactor test code and return the api error when needed
