+++ src/test/groovy/io/saagie/plugin/job/JobUpdateTaskTests.groovy
@@ -1,0 +1,175 @@
+package io.saagie.plugin.job
+
+import okhttp3.mockwebserver.MockResponse
+import okhttp3.mockwebserver.MockWebServer
+import org.gradle.testkit.runner.BuildResult
+import org.gradle.testkit.runner.GradleRunner
+import org.junit.Rule
+import org.junit.rules.TemporaryFolder
+import spock.lang.Shared
+import spock.lang.Specification
+import spock.lang.Title
+
+@Title('projectsUpdateJob task tests')
+class JobUpdateTaskTests extends Specification {
+    @Rule TemporaryFolder testProjectDir = new TemporaryFolder()
+    @Shared MockWebServer mockWebServer = new MockWebServer()
+
+    File buildFile
+    File jobFile
+
+    def setupSpec() {
+        mockWebServer.start(9000)
+    }
+
+    def cleanupSpec() {
+        mockWebServer.shutdown()
+    }
+
+    def setup() {
+        buildFile = testProjectDir.newFile('build.gradle')
+        buildFile << 'plugins { id "io.saagie.gradle-saagie-dataops-plugin" }\n'
+
+        jobFile = testProjectDir.newFile('jobFile.py')
+    }
+
+    def cleanup() {
+        mockWebServer.dispatcher.peek()
+    }
+
+    private BuildResult gradle(boolean isSuccessExpected, String[] arguments = ['tasks']) {
+        arguments += '--stacktrace'
+        def runner = GradleRunner.create()
+            .withArguments(arguments)
+            .withProjectDir(testProjectDir.root)
+            .withPluginClasspath()
+            .withDebug(true)
+
+        return isSuccessExpected ? runner.build() : runner.buildAndFail();
+    }
+
+    private BuildResult gradle(String[] arguments = ['tasks']) {
+        gradle(true, arguments)
+    }
+
+    def "projectsUpdateJob should update the specified job with only job config"() {
+        given:
+        def mockedJobCreationResponse = new MockResponse()
+        mockedJobCreationResponse.responseCode = 200
+        mockedJobCreationResponse.body = '''{"data":{"editJob":{"id":"jobId"}}}'''
+        mockWebServer.enqueue(mockedJobCreationResponse)
+
+        buildFile << '''
+            saagie {
+                server {
+                    url = 'http://localhost:9000'
+                    login = 'user.user'
+                    password = 'password'
+                    environment = 2
+                }
+                
+                project {
+                    id = 'projectId'
+                }
+                
+                job {
+                    id = 'jobId'
+                    name = 'Updated from gradle'
+                    description = 'updated description'
+                    alerting {
+                        emails = ['email@email.com']
+                        statusList = ['REQUESTED']
+                    }
+                }
+            }
+        '''
+
+        when:
+        BuildResult result = gradle 'projectsUpdateJob'
+
+        then:
+        result.output.contains('"id"')
+    }
+
+    def "projectsUpdateJob should fail if job id is missing"() {
+        given:
+        buildFile << '''
+            saagie {
+                server {
+                    url = 'http://localhost'
+                    login = 'user.user'
+                    password = 'password'
+                    environment = 2
+                }
+                
+                job {
+                    name = 'Updated from gradle'
+                    description = 'updated description'
+                    alerting {
+                        emails = ['email@email.com']
+                        statusList = ['REQUESTED']
+                    }
+                }
+            }
+        '''
+
+        when:
+        BuildResult result = gradle 'projectsUpdateJob'
+
+        then:
+        thrown(Exception)
+        result == null
+    }
+
+    def "projectsUpdateJob should add a new job version and upload script if config is provided"() {
+        given:
+        def mockedJobUpdateResponse = new MockResponse()
+        mockedJobUpdateResponse.responseCode = 200
+        mockedJobUpdateResponse.body = '''{"data":{"editJob":{"id":"jobId"}}}'''
+        mockWebServer.enqueue(mockedJobUpdateResponse)
+
+        def mockedJobVersionResponse = new MockResponse()
+        mockedJobVersionResponse.responseCode = 200
+        mockedJobVersionResponse.body = '''{"data":{"addJobVersion":{"number":"jobNumber"}}}'''
+        mockWebServer.enqueue(mockedJobVersionResponse)
+
+        def mockedJobFileUploadResponse = new MockResponse()
+        mockedJobFileUploadResponse.responseCode = 200
+        mockedJobFileUploadResponse.body = '''true'''
+        mockWebServer.enqueue(mockedJobFileUploadResponse)
+
+        buildFile << """
+            saagie {
+                server {
+                    url = 'http://localhost:9000'
+                    login = 'user.user'
+                    password = 'password'
+                    environment = 4
+                }
+                
+                project {
+                    id = 'projectId'
+                }
+                
+                job {
+                    id = 'jobId'
+                }
+                
+                jobVersion {
+                    commandLine = "python {file} 1 2 3"
+                    releaseNote = "Feat: won't fail"
+                    runtimeVersion = "3.6"
+                    packageInfo {
+                        name = "${jobFile.absolutePath}"
+                    }
+                }
+            }
+        """
+
+        when:
+        BuildResult result = gradle 'projectsUpdateJob'
+
+        then:
+        result.output.contains('"id"')
+    }
+}
