+++ src/test/groovy/io/saagie/plugin/project/ProjectListTaskTests.groovy
@@ -1,0 +1,113 @@
+package io.saagie.plugin.project
+
+import okhttp3.mockwebserver.MockResponse
+import okhttp3.mockwebserver.MockWebServer
+import org.gradle.testkit.runner.BuildResult
+import org.gradle.testkit.runner.GradleRunner
+import org.gradle.testkit.runner.UnexpectedBuildFailure
+import org.junit.Rule
+import org.junit.rules.TemporaryFolder
+import spock.lang.Shared
+import spock.lang.Specification
+import spock.lang.Title
+
+import static org.gradle.testkit.runner.TaskOutcome.FAILED
+
+@Title("projectsList task tests")
+class ProjectListTaskTests extends Specification {
+    @Rule TemporaryFolder testProjectDir = new TemporaryFolder()
+    @Shared MockWebServer mockWebServer = new MockWebServer()
+
+    File buildFile
+    File jobFile
+
+    def setupSpec() {
+        mockWebServer.start(9000)
+    }
+
+    def cleanupSpec() {
+        mockWebServer.shutdown()
+    }
+
+    def setup() {
+        buildFile = testProjectDir.newFile('build.gradle')
+        buildFile << 'plugins { id "io.saagie.gradle-saagie-dataops-plugin" }\n'
+
+        jobFile = testProjectDir.newFile('jobFile.py')
+    }
+
+    def cleanup() {
+        mockWebServer.dispatcher.peek()
+    }
+
+    private BuildResult gradle(boolean isSuccessExpected, String[] arguments = ['tasks']) {
+        arguments += '--stacktrace'
+        def runner = GradleRunner.create()
+            .withArguments(arguments)
+            .withProjectDir(testProjectDir.root)
+            .withPluginClasspath()
+            .withDebug(true)
+
+        return isSuccessExpected ? runner.build() : runner.buildAndFail();
+    }
+
+    private BuildResult gradle(String[] arguments = ['tasks']) {
+        gradle(true, arguments)
+    }
+
+    def "projectsList task should return a list of project"() {
+        given:
+        def mockedResponse = new MockResponse()
+        mockedResponse.responseCode = 200
+        mockedResponse.body = """
+            {"data":{"projects":[{"id":"8321e13c-892a-4481-8552-dekzdjeijzd","name":"Test new Project"},{"id":"7f5e0374-0c45-45a3-a2f3-dkjezoijdizd","name":"Test Spark config"},{"id":"bba3511b-7b7f-44f0-9f69-djeizjdoijzj","name":"For tests"},{"id":"9feae78d-1cc0-49bd-9e63-deozjiodjeiz","name":"Test simon"}]}}
+        """
+
+        buildFile << """
+            saagie {
+                server {
+                    url = 'http://localhost:9000'
+                    login = 'fake.user'
+                    password = 'ThisPasswordIsWrong'
+                    environment = 2
+                }
+            }
+        """
+        mockWebServer.enqueue(mockedResponse)
+
+        when:
+        def result = gradle 'projectsList'
+
+        then:
+        !result.output.contains('"data"')
+        result.output.contains('[{"id":"8321e13c')
+    }
+
+    def "projectsList task with bad config should fail"() {
+        given:
+        def mockedResponse = new MockResponse()
+        mockedResponse.responseCode = 400
+        mockWebServer.enqueue(mockedResponse)
+
+        buildFile << """
+            saagie {
+                server {
+                    url = 'http://localhost:9000'
+                    login = 'fake.user'
+                    password = 'ThisPasswordIsWrong'
+                    environment = 2
+                }
+            }
+        """
+
+        when:
+        BuildResult result = gradle 'projectsList'
+        println result?.output
+
+        then:
+        UnexpectedBuildFailure e = thrown()
+        e.message.contains('Error 400 when requesting on http://localhost:9000')
+        e.getBuildResult().task(':projectsList').outcome == FAILED
+    }
+
+}
